auth-service/requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0
PyJWT==2.8.0
passlib==1.7.4
